pyproject.toml
[tool.poetry]
name = "hbtnet"
version = "0.1.0"
description = ""
authors = ["7cm-diameter <from.this.ridiculous.loop@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8,<3.12"
numpy = "^1.23.2"
scipy = "^1.9.0"
networkx = "^2.8.5"
pandas = "^1.4.3"
PyYAML = "^6.0"
argparse = "^1.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-language-server = {extras = ["all"], version = "^0.36.2"}
isort = "^5.10.1"
yapf = "^0.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
